day6/agent-communication-security/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
google-generativeai==0.3.2
websockets==12.0
pytest==7.4.4
pytest-asyncio==0.23.2
python-multipart==0.0.6
aiofiles==23.2.1
structlog==23.2.0
prometheus-client==0.19.0
